Refactored file names and re-arranged the folders
