+++ src/components/detailedView/DetailedViewerContainer.tsx
