+++ src/components/detailedView/displays/index.ts
