+++ src/components/detailedView/viewer.scss
