+++ src/components/feed/Preview/FileDetailView.tsx
@@ -4,8 +4,11 @@ import { ErrorBoundary } from "react-error-boundary";
 
 import { FeedFile } from "@fnndsc/chrisapi";
 import { ExpandIcon, FilmIcon, InfoCircleIcon } from "@patternfly/react-icons";
-import { getFileExtension } from "../../api/models/file-explorer.model";
-import { IFileBlob, fileViewerMap } from "../../api/models/file-viewer.model";
+import { getFileExtension } from "../../../api/models/file-explorer.model";
+import {
+  IFileBlob,
+  fileViewerMap,
+} from "../../../api/models/file-viewer.model";
 
 const ViewerDisplay = React.lazy(() => import("./displays/ViewerDisplay"));
