+++ src/components/feed/Preview/displays/IframeDisplay.tsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from "react";
-import { IFileBlob } from "../../../api/models/file-viewer.model";
+import { IFileBlob } from "../../../../api/models/file-viewer.model";
 type AllProps = {
   fileItem: IFileBlob;
 };
