+++ src/components/feed/Preview/displays/ImageDisplay.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { IFileBlob } from "../../../api/models/file-viewer.model";
+import { IFileBlob } from "../../../../api/models/file-viewer.model";
 
 type AllProps = {
   fileItem: IFileBlob;
