+++ src/components/feed/Preview/displays/ViewerDisplay.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { IFileBlob } from "../../../api/models/file-viewer.model";
+import { IFileBlob } from "../../../../api/models/file-viewer.model";
 import {
   CatchallDisplay,
   JsonDisplay,
