+++ src/components/gallery/GalleryWrapper.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { IGalleryToolbarState } from "../../store/gallery/types";
 import { galleryActions } from "../../api/models/gallery.model";
-import { GalleryToolbar } from "../gallery";
+import { GalleryToolbar } from ".";
 import "./GalleryWrapper.scss";
 
 type AllProps = {
